--- automerge-backend/src/helper.rs
@@ -1,3 +1,0 @@
-pub(crate) fn make_true() -> bool {
-    true
-}
