+++ automerge-backend/src/lib.rs
@@ -19,22 +19,18 @@ mod columnar;
 mod concurrent_operations;
 mod encoding;
 mod error;
-mod helper;
 mod object_store;
 mod op_handle;
 mod op_set;
 mod ordered_set;
 mod patch;
-mod protocol;
+mod undo_operation;
 mod serialize;
 mod time;
 mod actor_map;
 
-pub use automerge_protocol::{ActorID, ChangeHash, ObjType, ObjectID, ElementID, Key, DataType, Value, OpType, Operation};
+pub use automerge_protocol::{ActorID, ChangeHash, ObjType, ObjectID, ElementID, Key, DataType, Value, OpType, Operation, Change, ChangeRequest, ChangeRequestType,};
 
 pub use crate::patch::{DiffEdit, Patch, Diff, MapDiff, SeqDiff, ObjDiff};
-pub use crate::protocol::{
-    Change, ChangeRequest, ChangeRequestType,
-};
 pub use backend::Backend;
 pub use error::AutomergeError;
