+++ automerge-protocol/src/lib.rs
@@ -294,3 +294,48 @@ impl Operation {
     }
 }
 
+#[derive(Deserialize, Serialize, PartialEq, Debug, Clone)]
+pub struct Change {
+    #[serde(rename = "ops")]
+    pub operations: Vec<Operation>,
+    #[serde(rename = "actor")]
+    pub actor_id: ActorID,
+    pub hash: ChangeHash,
+    pub seq: u64,
+    #[serde(rename = "startOp")]
+    pub start_op: u64,
+    pub time: i64,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub message: Option<String>,
+    pub deps: Vec<ChangeHash>,
+}
+
+impl Change {
+    pub fn max_op(&self) -> u64 {
+        self.start_op + (self.operations.len() as u64) - 1
+    }
+}
+
+#[derive(Deserialize, PartialEq, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct ChangeRequest {
+    pub actor: ActorID,
+    pub seq: u64,
+    pub version: u64,
+    pub message: Option<String>,
+    #[serde(default = "serde_impls::make_true")]
+    pub undoable: bool,
+    pub time: Option<i64>,
+    pub deps: Option<Vec<ChangeHash>>,
+    pub ops: Option<Vec<OpRequest>>,
+    pub child: Option<String>,
+    pub request_type: ChangeRequestType,
+}
+
+#[derive(Deserialize, PartialEq, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub enum ChangeRequestType {
+    Change,
+    Undo,
+    Redo,
+}
