+++ automerge-protocol/src/serde_impls/mod.rs
@@ -11,3 +11,7 @@ mod operation;
 pub(crate) fn make_false() -> bool {
     return false
 }
+
+pub(crate) fn make_true() -> bool {
+    return false
+}
